Split project into separate Maven modules
