+++ blockchain-agent-proof/pom.xml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>com.sphereon.alfresco.blockchain-agent</groupId>
+        <artifactId>blockchain-agent-parent</artifactId>
+        <version>0.0.2-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>blockchain-agent-proof</artifactId>
+
+    <packaging>jar</packaging>
+
+    <properties>
+        <blockchain-proof-sdk-java8.version>0.10.0</blockchain-proof-sdk-java8.version>
+        <easy-blockchain-lib.version>0.1.3</easy-blockchain-lib.version>
+        <crypto-keys-sdk-java8-okhttp-gson.version>0.9.0</crypto-keys-sdk-java8-okhttp-gson.version>
+        <authentication-lib.version>0.1.4-SNAPSHOT</authentication-lib.version>
+        <sphereon-commons-bom.version>1.0.4-SNAPSHOT</sphereon-commons-bom.version>
+    </properties>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <mainClass>com.sphereon.alfresco.blockchain.agent.AlfrescoBlockchainAgentApp</mainClass>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+
+    <dependencies>
+        <dependency>
+            <groupId>${project.parent.groupId}</groupId>
+            <artifactId>blockchain-agent</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.sphereon.public</groupId>
+            <artifactId>authentication-lib-main</artifactId>
+            <version>${authentication-lib.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.sphereon.sdk</groupId>
+            <artifactId>blockchain-proof-sdk-java8-okhttp-gson</artifactId>
+            <version>${blockchain-proof-sdk-java8.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.sphereon.public</groupId>
+            <artifactId>easy-blockchain-lib-main</artifactId>
+            <version>${easy-blockchain-lib.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.sphereon.sdk</groupId>
+            <artifactId>crypto-keys-sdk-java8-okhttp-gson</artifactId>
+            <version>${crypto-keys-sdk-java8-okhttp-gson.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+</project>
