+++ blockchain-agent-proof/src/main/java/com/sphereon/alfresco/blockchain/agent/proof/tasks/ProofVerifyPendingRegistrationsTask.java
