+++ blockchain-agent/src/main/java/com/sphereon/alfresco/blockchain/agent/config/AppConfig.java
