+++ blockchain-agent/src/main/java/com/sphereon/alfresco/blockchain/agent/model/AlfrescoBlockchainRegistrationState.java
