+++ blockchain-agent/src/main/java/com/sphereon/alfresco/blockchain/agent/rest/model/ErrorResponse.java
