+++ blockchain-agent/src/main/java/com/sphereon/alfresco/blockchain/agent/tasks/ondemand/VerifyRegistrationTask.java
