+++ blockchain-agent/src/main/java/com/sphereon/alfresco/blockchain/agent/utils/Signer.java
