+++ blockchain-agent/src/main/resources/log4j2.xml
